feat: app state, sound hit and miss hit
