+++ src/components/BCell/BCell.scss
@@ -2,6 +2,7 @@
   width: 30%;
   height: auto;
   position: relative;
+  overflow: hidden;
 }
 
 .square:after {
@@ -16,7 +17,7 @@
   bottom: 0;
   left: 0;
   right: 0;
-  background: gold;
+  background: transparent;
   padding: 0;
   width: 100%;
   height: 100;
