+++ src/components/See/See.jsx
@@ -1,8 +1,13 @@
 import * as React from 'react';
+import { useEffect } from 'react';
 import { useSelector, useDispatch } from 'react-redux';
 import PropTypes from 'prop-types';
 import './See.scss';
 import { Button } from 'react-bootstrap';
+
+// Sound
+import { Howl } from 'howler';
+import laser from '~audio/laser.mp3';
 
 import { HERE_IS_LOSER, ATTACK_TIME } from '~constants';
 import { missHit, countAttacks, randomPlay } from '~store/game/actions';
@@ -13,6 +18,7 @@ export default function See({
   const classes = `content see ${value === HERE_IS_LOSER ? 'lose' : ''}`;
   const dispatch = useDispatch();
   const whoseTurn = useSelector(({ game: { activePlayer } }) => activePlayer);
+  const userBoard = useSelector(({ game: { user } }) => user);
   const enemy = +!whoseTurn;
   const boardState = useSelector(({ game: { [whoseTurn]: { rows } } }) => rows);
   const boardSize = useSelector(({ game: { size } }) => size);
@@ -21,6 +27,11 @@ export default function See({
   const isAutoPlay = useSelector(({ game: { [userTurn]: { autoPlay } } }) => autoPlay);
 
   function onAttack() {
+    // Sound
+    const sound = new Howl({
+      src: [laser],
+    });
+    sound.play();
     dispatch(countAttacks(whoseTurn));
     dispatch(missHit(id, num, player));
     const interval = setTimeout(
@@ -28,6 +39,16 @@ export default function See({
     );
     return () => clearInterval(interval);
   }
+
+  useEffect(() => {
+    if (player === userBoard && value === HERE_IS_LOSER) {
+      // Sound
+      const sound = new Howl({
+        src: [laser],
+      });
+      sound.play();
+    }
+  }, [value, player, userBoard, dispatch]);
 
   return (
     <Button
