+++ src/components/Ship/Ship.jsx
@@ -5,6 +5,11 @@ import PropTypes from 'prop-types';
 import './Ship.scss';
 import { Button } from 'react-bootstrap';
 
+// Sound
+import { Howl } from 'howler';
+import shipExplosion from '~audio/ship-explosion.wav';
+import laser from '~audio/laser.mp3';
+
 import { HERE_IS_FIRE } from '~constants';
 import {
   getHit, countAttacks, countFired, resetLife,
@@ -19,11 +24,20 @@ export default function Ship({
   const userBoard = useSelector(({ game: { user } }) => user);
   const whoseTurn = useSelector(({ game: { activePlayer } }) => activePlayer);
   const classes = `content ship ${value === HERE_IS_FIRE ? 'fired' : ''}`;
-  const colorShip = useSelector(({ game: { shipColor } }) => shipColor);
+  const colorShip = useSelector(({ app: { shipColor } }) => shipColor);
   const isAutoPlay = useSelector(({ game: { [userBoard]: { autoPlay } } }) => autoPlay);
   const fired = useSelector(({ game: { [player]: { firedShips } } }) => firedShips);
 
   function onAttack() {
+    // Sound
+    const sound1 = new Howl({
+      src: [laser],
+    });
+    sound1.play();
+    const sound2 = new Howl({
+      src: [shipExplosion],
+    });
+    sound2.play();
     dispatch(countAttacks(whoseTurn));
     dispatch(countFired(player));
     dispatch(resetLife({ player, fired }));
@@ -32,8 +46,15 @@ export default function Ship({
 
   useEffect(() => {
     if (player === userBoard && value === HERE_IS_FIRE) {
-      dispatch(countFired(player));
-      dispatch(resetLife({ player, fired }));
+      // Sound
+      const sound1 = new Howl({
+        src: [laser],
+      });
+      sound1.play();
+      const sound2 = new Howl({
+        src: [shipExplosion],
+      });
+      sound2.play();
     }
   }, [value, player, userBoard, fired, dispatch]);
