+++ src/components/Ship/svg/ShipSVG.scss
@@ -1,5 +1,4 @@
 .ship-svg {
-  transform: scale(2);
   &.hidden {
     opacity: 0;
   }
