+++ src/index.scss
@@ -3,17 +3,8 @@
 body {
   margin: 0;
   box-sizing: border-box;
-  font-family: -apple-system,
- BlinkMacSystemFont,
- "Segoe UI",
- Roboto,
- Oxygen,
- Ubuntu,
- Cantarell,
-    "Fira Sans",
- "Droid Sans",
- "Helvetica Neue",
- sans-serif;
+  font-family: -apple-system, BlinkMacSystemFont, "Segoe UI", Roboto, Oxygen, Ubuntu, Cantarell,
+    "Fira Sans", "Droid Sans", "Helvetica Neue", sans-serif;
   -webkit-font-smoothing: antialiased;
   -moz-osx-font-smoothing: grayscale;
 }
@@ -21,6 +12,7 @@ body {
 .root {
   min-height: 100vh;
   color: #ed044b;
+  overflow: hidden;
 }
 
 .btn.btn-primary {
