+++ src/store/app/action-types.js
@@ -1,0 +1,20 @@
+// WARNINGS
+export const ALERT_MESSAGE = 'APP/ALERT_MESSAGE';
+export const CLEAR_ALERT_MESSAGE = 'APP/CLEAR_ALERT_MESSAGE';
+
+// Loaders
+export const HIDE_LOADER = 'APP/HIDE_LOADER';
+export const SHOW_LOADER = 'APP/SHOW_LOADER';
+
+// Sound & Music
+export const RESET_SOUND = 'APP/RESET_SOUND';
+
+// Modals
+export const TOGGLE_WELCOME_MODAL = 'APP/TOGGLE_WELCOME_MODAL';
+export const TOGGLE_RECORDS_MODAL = 'APP/TOGGLE_RECORDS_MODAL';
+export const TOGGLE_FINISH_MODAL = 'APP/TOGGLE_FINISH_MODAL';
+export const TOGGLE_OPTIONS_MODAL = 'APP/TOGGLE_OPTIONS_MODAL';
+
+// Appearance
+export const REFRESH_BACKGROUND = 'APP/REFRESH_BACKGROUND';
+export const REFRESH_SHIP_COLOR = 'APP/REFRESH_SHIP_COLOR';
