+++ src/store/game/action-types.js
@@ -1,0 +1,30 @@
+// STORAGE
+export const WRITE_SUCCESS = 'STORAGE/WRITE_SUCCESS';
+export const WRITE_FAILURE = 'STORAGE/WRITE_FAILURE';
+export const READ_SUCCESS = 'STORAGE/READ_SUCCESS';
+export const READ_FAILURE = 'STORAGE/READ_FAILURE';
+
+// GAME
+export const RESET_IS_PLAYING = 'GAME/RESET_IS_PLAYING';
+export const SAVE_TO_RECORDS = 'GAME/SAVE_TO_RECORDS';
+export const GAME_OVER = 'GAME/GAME_OVER';
+export const RESET_AUTO_PLAY = 'GAME/RESET_AUTO_PLAY';
+export const RESET_ACTIVE_PLAYER = 'GAME/RESET_ACTIVE_PLAYER';
+export const RESET_PROGRESS = 'GAME/RESET_PROGRESS';
+export const RESET_SHIP_NUM = 'GAME/RESET_SHIP_NUM';
+export const RESET_VALID_SHIP_NUM = 'GAME/RESET_VALID_SHIP_NUM';
+export const SET_GAME_STATE = 'GAME/SET_GAME_STATE';
+
+// BOARD
+export const SET_RANDOM_SHIP_POSITIONS = 'BOARD/SET_RANDOM_SHIP_POSITIONS';
+
+// PLAYER
+export const SET_PLAYER_STATE = 'PLAYER/SET_PLAYER_STATE';
+export const RESET_LIFE = 'PLAYER/RESET_LIFE';
+
+// ATTACKS
+export const HIT_SHIP = 'ATTACK/HIT_SHIP';
+export const MISS_HIT = 'ATTACK/MISS_HIT';
+export const COUNT_ATTACK = 'ATTACK/COUNT_ATTACK';
+export const COUNT_FIRED_SHIPS = 'ATTACK/COUNT_FIRED_SHIPS';
+export const AUTO_ATTACK = 'ATTACK/AUTO_ATTACK';
