+++ src/store/reducer.js
@@ -1,10 +1,11 @@
 import { combineReducers } from 'redux';
 
 import gameReducer from '~store/game/reducer';
+import appReducer from '~store/app/reducer';
 
 const rootReducer = combineReducers({
   game: gameReducer,
-  // user: userReducer,
+  app: appReducer,
 });
 
 export default rootReducer;
